4/5/20

+++ PHP/Laravel/Examples/#1/database/migrations/2020_05_01_144745_create_blog_categories_table.php
@@ -1,11 +1,11 @@
 <?php
 
-use Illuminate\Database\Migrations\Migration;
-use Illuminate\Database\Schema\Blueprint;
-use Illuminate\Support\Facades\Schema;
+  use Illuminate\Database\Migrations\Migration;
+  use Illuminate\Database\Schema\Blueprint;
+  use Illuminate\Support\Facades\Schema;
 
-class CreateBlogCategoriesTable extends Migration
-{
+  class CreateBlogCategoriesTable extends Migration
+  {
     /**
      * Run the migrations.
      *
@@ -13,17 +13,17 @@ class CreateBlogCategoriesTable extends Migration
      */
     public function up()
     {
-        Schema::create('blog_categories', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('parent_id')->unsigned()->default(1); // unsigned >= 0
+      Schema::create('blog_categories', function (Blueprint $table) {
+        $table->bigIncrements('id');
+        $table->bigInteger('parent_id')->unsigned()->default(1); // unsigned >= 0
 
-            $table->string('slug')->unique();
-            $table->string('title');
-            $table->text('description')->nullable(); // nullable - необязательное
+        $table->string('slug')->unique();
+        $table->string('title');
+        $table->text('description')->nullable(); // nullable - необязательное
 
-            $table->timestamps(); // CREATED AT + UPDATED AD
-            $table->softDeletes(); // при удалении остается запись в БД (в модели прописать trait use SoftDeletes)
-        });
+        $table->timestamps(); // CREATED AT + UPDATED AD
+        $table->softDeletes(); // при удалении остается запись в БД (в модели прописать trait use SoftDeletes)
+      });
     }
 
     /**
@@ -33,6 +33,6 @@ class CreateBlogCategoriesTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('blog_categories');
+      Schema::dropIfExists('blog_categories');
     }
-}
+  }
